Update Game content

+++ plugins/core/client/src/Madgine/render/rendertarget.cpp
@@ -65,7 +65,7 @@ namespace Render {
                 { { 1, 1, 0 } }
             };
 
-            std::vector<unsigned short> indices {
+            std::vector<uint32_t> indices {
                 0, 1, 2, 1, 2, 3
             };
