+++ plugins/renderer/directx12/src/DirectX12/directx12rendertarget.cpp
@@ -246,7 +246,7 @@ namespace Render {
         }
     }
 
-    void DirectX12RenderTarget::renderVertices(const Program *program, size_t groupSize, std::vector<Vertex> vertices, std::vector<unsigned short> indices)
+    void DirectX12RenderTarget::renderVertices(const Program *program, size_t groupSize, std::vector<Vertex> vertices, std::vector<uint32_t> indices)
     {
         if (!vertices.empty()) {
             DirectX12MeshData tempMesh;
@@ -258,7 +258,7 @@ namespace Render {
         }
     }
 
-    void DirectX12RenderTarget::renderVertices(const Program *program, size_t groupSize, std::vector<Vertex2> vertices, std::vector<unsigned short> indices, const Material *material)
+    void DirectX12RenderTarget::renderVertices(const Program *program, size_t groupSize, std::vector<Vertex2> vertices, std::vector<uint32_t> indices, const Material *material)
     {
         if (!vertices.empty()) {
             DirectX12MeshData tempMesh;
