+++ plugins/renderer/opengl/src/OpenGL/openglrendertarget.h
@@ -16,8 +16,8 @@ namespace Render {
         virtual void popAnnotation() override;
 
         virtual void setRenderSpace(const Rect2i &space) override;
-        virtual void renderVertices(const Program *program, size_t groupSize, std::vector<Vertex> vertices, std::vector<unsigned short> indices = {}) override;
-        virtual void renderVertices(const Program *program, size_t groupSize, std::vector<Vertex2> vertices, std::vector<unsigned short> indices = {}, const Material *material = nullptr) override;
+        virtual void renderVertices(const Program *program, size_t groupSize, std::vector<Vertex> vertices, std::vector<uint32_t> indices = {}) override;
+        virtual void renderVertices(const Program *program, size_t groupSize, std::vector<Vertex2> vertices, std::vector<uint32_t> indices = {}, const Material *material = nullptr) override;
         virtual void renderMesh(const GPUMeshData *mesh, const Program *program, const Material *material = nullptr) override;
         virtual void renderMeshInstanced(size_t count, const GPUMeshData *mesh, const Program *program, const Material *material = nullptr) override;
         virtual void clearDepthBuffer() override;
