+++ shared/im3d/im3d_internal.h
@@ -29,11 +29,11 @@ namespace Im3D {
 
         struct RenderData {
             std::vector<Render::Vertex> mVertices[IM3D_MESHTYPE_COUNT];
-            std::vector<unsigned short> mIndices[IM3D_MESHTYPE_COUNT];
+            std::vector<uint32_t> mIndices[IM3D_MESHTYPE_COUNT];
             size_t mVertexBase[IM3D_MESHTYPE_COUNT];
             std::list<std::tuple<std::chrono::steady_clock::time_point, std::vector<Render::Vertex>, std::vector<unsigned short>>> mPersistentMeshes[IM3D_MESHTYPE_COUNT];
             std::vector<Render::Vertex2> mVertices2[IM3D_MESHTYPE_COUNT];
-            std::vector<unsigned short> mIndices2[IM3D_MESHTYPE_COUNT];
+            std::vector<uint32_t> mIndices2[IM3D_MESHTYPE_COUNT];
             size_t mVertexBase2[IM3D_MESHTYPE_COUNT];
             Render::RenderPassFlags mFlags = Render::RenderPassFlags_NoLighting;
         };
